+++ R/class_mermaid.R
@@ -2,6 +2,7 @@ mermaid_init <- function(
   network,
   label = NULL,
   label_break = "<br>",
+  label_length = 60L,
   label_width = 30L,
   show_legend = TRUE,
   show_color = TRUE
@@ -10,6 +11,7 @@ mermaid_init <- function(
     network = network,
     label = label,
     label_break = label_break,
+    label_length = label_length,
     label_width = label_width,
     show_legend = show_legend,
     show_color = show_color
@@ -20,6 +22,7 @@ mermaid_new <- function(
   network = NULL,
   label = NULL,
   label_break = NULL,
+  label_length = NULL,
   label_width = NULL,
   show_legend = NULL,
   show_color = NULL
@@ -28,6 +31,7 @@ mermaid_new <- function(
     network = network,
     label = label,
     label_break = label_break,
+    label_length = label_length,
     label_width = label_width,
     show_legend = show_legend,
     show_color = show_color
@@ -47,6 +51,7 @@ mermaid_class <- R6::R6Class(
       network = NULL,
       label = NULL,
       label_break = NULL,
+      label_length = NULL,
       label_width = NULL,
       show_legend = NULL,
       show_color = NULL
@@ -55,6 +60,7 @@ mermaid_class <- R6::R6Class(
         network = network,
         label = label,
         label_break = label_break,
+        label_length = label_length,
         label_width = label_width
       )
       self$show_legend <- show_legend