+++ R/tar_glimpse.R
@@ -17,8 +17,10 @@
 #' @param label Character vector of one or more aesthetics to add to the
 #'   vertex labels. Currently, the only option is  `"description"` to show each
 #'   target's custom description, or `character(0)` to suppress it.
+#' @param label_length Positive numeric of length 1, maximum
+#'   number of content characters of each node label.
 #' @param label_width Positive numeric of length 1, maximum width
-#'   (in number of characters) of the node labels.
+#'   of each node label.
 #' @param level_separation Numeric of length 1,
 #'   `levelSeparation` argument of `visNetwork::visHierarchicalLayout()`.
 #'   Controls the distance between hierarchical levels.
@@ -61,6 +63,7 @@ tar_glimpse <- function(
   allow = NULL,
   exclude = ".Random.seed",
   label = targets::tar_config_get("label"),
+  label_length = targets::tar_config_get("label_length"),
   label_width = targets::tar_config_get("label_width"),
   level_separation = targets::tar_config_get("level_separation"),
   degree_from = 1L,
@@ -78,6 +81,9 @@ tar_glimpse <- function(
   tar_assert_package("visNetwork")
   tar_assert_lgl(targets_only)
   tar_assert_in(label, "description")
+  tar_assert_dbl(label_length)
+  tar_assert_scalar(label_length)
+  tar_assert_none_na(label_length)
   tar_assert_dbl(label_width)
   tar_assert_scalar(label_width)
   tar_assert_none_na(label_width)
@@ -102,6 +108,7 @@ tar_glimpse <- function(
     allow_quosure = rlang::enquo(allow),
     exclude_quosure = rlang::enquo(exclude),
     label = label,
+    label_length = label_length,
     label_width = label_width,
     level_separation = level_separation,
     degree_from = degree_from,
@@ -130,6 +137,7 @@ tar_glimpse_inner <- function(
   allow_quosure,
   exclude_quosure,
   label,
+  label_length,
   label_width,
   level_separation,
   degree_from,
@@ -153,6 +161,7 @@ tar_glimpse_inner <- function(
   visual <- visnetwork_init(
     network = network,
     label = label,
+    label_length = label_length,
     label_width = label_width,
     level_separation = level_separation,
     degree_from = degree_from,